Replace `checker.Vuln` with `clients.Vuln`
